requirements.txt
aioredis==1.3.1
aiosqlite==0.17.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.1
daphne==3.0.2
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.46.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iso8601==1.0.2
msgpack==1.0.3
oauthlib==3.2.0
Pillow==9.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pypika-tortoise==0.1.4
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tortoise-orm==0.19.0
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
urllib3==1.26.9
zope.interface==5.4.0
